:recycle: Refactor: team registration test code
  - 팀 등록 위해 작성된 코드 리펙터링 진행
