+++ crispin/src/main/java/study/crispin/team/application/request/TeamRegistrationRequest.java
@@ -1,0 +1,4 @@
+package study.crispin.team.application.request;
+
+public record TeamRegistrationRequest(String name) {
+}
