+++ crispin/src/main/java/study/crispin/team/presentation/response/TeamRegistrationResponse.java
@@ -1,0 +1,4 @@
+package study.crispin.team.presentation.response;
+
+public record TeamRegistrationResponse(Long id, String name, String manager) {
+}
